requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
sqlparse==0.5.5
whitenoise==6.11.0
gunicorn==23.0.0
packaging==25.0
python-dotenv==1.2.1
Pillow==25.3
stripe>=8.0.0